# Holds most fields as slice of values, not ptrs in wasm.Module