--- a/internal/integration_test/fuzz/wazerolib/nodiff.go
+++ b/internal/integration_test/fuzz/wazerolib/nodiff.go
@@ -126,8 +126,7 @@ func requireNoDiff(wasmBin []byte, checkMemory bool, requireNoError func(err err
 // ensureDummyImports instantiates the modules which are required imports by `origin` *wasm.Module.
 func ensureDummyImports(r wazero.Runtime, origin *wasm.Module, requireNoError func(err error)) (skip bool) {
 	impMods := make(map[string][]wasm.Import)
-	for i := range origin.ImportSection {
-		imp := &origin.ImportSection[i]
+	for _, imp := range origin.ImportSection {
 		impMods[imp.Module] = append(impMods[imp.Module], imp)
 	}