--- a/internal/testing/binaryencoding/const_expr.go
+++ b/internal/testing/binaryencoding/const_expr.go
@@ -4,7 +4,7 @@ import (
 	"github.com/tetratelabs/wazero/internal/wasm"
 )
 
-func encodeConstantExpression(expr *wasm.ConstantExpression) (ret []byte) {
+func encodeConstantExpression(expr wasm.ConstantExpression) (ret []byte) {
 	ret = append(ret, expr.Opcode)
 	ret = append(ret, expr.Data...)
 	ret = append(ret, wasm.OpcodeEnd)