--- a/internal/testing/binaryencoding/data.go
+++ b/internal/testing/binaryencoding/data.go
@@ -8,7 +8,7 @@ import (
 func encodeDataSegment(d *wasm.DataSegment) (ret []byte) {
 	// Currently multiple memories are not supported.
 	ret = append(ret, leb128.EncodeInt32(0)...)
-	ret = append(ret, encodeConstantExpression(*d.OffsetExpression)...)
+	ret = append(ret, encodeConstantExpression(d.OffsetExpression)...)
 	ret = append(ret, leb128.EncodeUint32(uint32(len(d.Init)))...)
 	ret = append(ret, d.Init...)
 	return