--- a/internal/testing/binaryencoding/global.go
+++ b/internal/testing/binaryencoding/global.go
@@ -7,7 +7,7 @@ import (
 // encodeGlobal returns the wasm.Global encoded in WebAssembly 1.0 (20191205) Binary Format.
 //
 // See https://www.w3.org/TR/2019/REC-wasm-core-1-20191205/#global-section%E2%91%A0
-func encodeGlobal(g *wasm.Global) (data []byte) {
+func encodeGlobal(g wasm.Global) (data []byte) {
 	var mutable byte
 	if g.Type.Mutable {
 		mutable = 1