--- a/internal/wasm/binary/const_expr.go
+++ b/internal/wasm/binary/const_expr.go
@@ -11,10 +11,10 @@ import (
 	"github.com/tetratelabs/wazero/internal/wasm"
 )
 
-func decodeConstantExpression(r *bytes.Reader, enabledFeatures api.CoreFeatures) (wasm.ConstantExpression, error) {
+func decodeConstantExpression(r *bytes.Reader, enabledFeatures api.CoreFeatures, ret *wasm.ConstantExpression) error {
 	b, err := r.ReadByte()
 	if err != nil {
-		return wasm.ConstantExpression{}, fmt.Errorf("read opcode: %v", err)
+		return fmt.Errorf("read opcode: %v", err)
 	}
 
 	remainingBeforeData := int64(r.Len())
@@ -31,44 +31,44 @@ func decodeConstantExpression(r *bytes.Reader, enabledFeatures api.CoreFeatures)
 	case wasm.OpcodeF32Const:
 		buf := make([]byte, 4)
 		if _, err := io.ReadFull(r, buf); err != nil {
-			return wasm.ConstantExpression{}, fmt.Errorf("read f32 constant: %v", err)
+			return fmt.Errorf("read f32 constant: %v", err)
 		}
 		_, err = ieee754.DecodeFloat32(buf)
 	case wasm.OpcodeF64Const:
 		buf := make([]byte, 8)
 		if _, err := io.ReadFull(r, buf); err != nil {
-			return wasm.ConstantExpression{}, fmt.Errorf("read f64 constant: %v", err)
+			return fmt.Errorf("read f64 constant: %v", err)
 		}
 		_, err = ieee754.DecodeFloat64(buf)
 	case wasm.OpcodeGlobalGet:
 		_, _, err = leb128.DecodeUint32(r)
 	case wasm.OpcodeRefNull:
 		if err := enabledFeatures.RequireEnabled(api.CoreFeatureBulkMemoryOperations); err != nil {
-			return wasm.ConstantExpression{}, fmt.Errorf("ref.null is not supported as %w", err)
+			return fmt.Errorf("ref.null is not supported as %w", err)
 		}
 		reftype, err := r.ReadByte()
 		if err != nil {
-			return wasm.ConstantExpression{}, fmt.Errorf("read reference type for ref.null: %w", err)
+			return fmt.Errorf("read reference type for ref.null: %w", err)
 		} else if reftype != wasm.RefTypeFuncref && reftype != wasm.RefTypeExternref {
-			return wasm.ConstantExpression{}, fmt.Errorf("invalid type for ref.null: 0x%x", reftype)
+			return fmt.Errorf("invalid type for ref.null: 0x%x", reftype)
 		}
 	case wasm.OpcodeRefFunc:
 		if err := enabledFeatures.RequireEnabled(api.CoreFeatureBulkMemoryOperations); err != nil {
-			return wasm.ConstantExpression{}, fmt.Errorf("ref.func is not supported as %w", err)
+			return fmt.Errorf("ref.func is not supported as %w", err)
 		}
 		// Parsing index.
 		_, _, err = leb128.DecodeUint32(r)
 	case wasm.OpcodeVecPrefix:
 		if err := enabledFeatures.RequireEnabled(api.CoreFeatureSIMD); err != nil {
-			return wasm.ConstantExpression{}, fmt.Errorf("vector instructions are not supported as %w", err)
+			return fmt.Errorf("vector instructions are not supported as %w", err)
 		}
 		opcode, err = r.ReadByte()
 		if err != nil {
-			return wasm.ConstantExpression{}, fmt.Errorf("read vector instruction opcode suffix: %w", err)
+			return fmt.Errorf("read vector instruction opcode suffix: %w", err)
 		}
 
 		if opcode != wasm.OpcodeVecV128Const {
-			return wasm.ConstantExpression{}, fmt.Errorf("invalid vector opcode for const expression: %#x", opcode)
+			return fmt.Errorf("invalid vector opcode for const expression: %#x", opcode)
 		}
 
 		remainingBeforeData = int64(r.Len())
@@ -76,30 +76,30 @@ func decodeConstantExpression(r *bytes.Reader, enabledFeatures api.CoreFeatures)
 
 		n, err := r.Read(make([]byte, 16))
 		if err != nil {
-			return wasm.ConstantExpression{}, fmt.Errorf("read vector const instruction immediates: %w", err)
+			return fmt.Errorf("read vector const instruction immediates: %w", err)
 		} else if n != 16 {
-			return wasm.ConstantExpression{}, fmt.Errorf("read vector const instruction immediates: needs 16 bytes but was %d bytes", n)
+			return fmt.Errorf("read vector const instruction immediates: needs 16 bytes but was %d bytes", n)
 		}
 	default:
-		return wasm.ConstantExpression{}, fmt.Errorf("%v for const expression opt code: %#x", ErrInvalidByte, b)
+		return fmt.Errorf("%v for const expression opt code: %#x", ErrInvalidByte, b)
 	}
 
 	if err != nil {
-		return wasm.ConstantExpression{}, fmt.Errorf("read value: %v", err)
+		return fmt.Errorf("read value: %v", err)
 	}
 
 	if b, err = r.ReadByte(); err != nil {
-		return wasm.ConstantExpression{}, fmt.Errorf("look for end opcode: %v", err)
+		return fmt.Errorf("look for end opcode: %v", err)
 	}
 
 	if b != wasm.OpcodeEnd {
-		return wasm.ConstantExpression{}, fmt.Errorf("constant expression has been not terminated")
+		return fmt.Errorf("constant expression has been not terminated")
 	}
 
-	data := make([]byte, remainingBeforeData-int64(r.Len())-1)
-	if _, err := r.ReadAt(data, offsetAtData); err != nil {
-		return wasm.ConstantExpression{}, fmt.Errorf("error re-buffering ConstantExpression.Data")
+	ret.Data = make([]byte, remainingBeforeData-int64(r.Len())-1)
+	if _, err = r.ReadAt(ret.Data, offsetAtData); err != nil {
+		return fmt.Errorf("error re-buffering ConstantExpression.Data")
 	}
-
-	return wasm.ConstantExpression{Opcode: opcode, Data: data}, nil
+	ret.Opcode = opcode
+	return nil
 }