… builder_test.go
… experimental/listener_test.go
… internal/engine/compiler/compiler_global_test.go
… internal/engine/compiler/engine_test.go
… internal/integration_test/bench/bench_test.go
… internal/integration_test/bench/hostfunc_bench_test.go
… internal/integration_test/engine/adhoc_test.go
… internal/integration_test/fuzzcases/fuzzcases_test.go
… internal/testing/binaryencoding/encoder_test.go
… internal/testing/binaryencoding/global_test.go
… internal/testing/binaryencoding/import_test.go
… internal/wasm/binary/const_expr_test.go
… internal/wasm/binary/data_test.go
… internal/wasm/binary/decoder_test.go
… internal/wasm/binary/element_test.go
… internal/wasm/binary/export.go
… internal/wasm/binary/global.go
… internal/wasm/binary/import.go
… internal/wasm/binary/import_test.go
… internal/wasm/binary/section.go
… internal/wasm/binary/section_test.go
… internal/wasm/binary/table.go
… internal/wasm/binary/table_test.go
… internal/wasm/counts.go
… internal/wasm/counts_test.go
… internal/wasm/func_validation.go
… internal/wasm/func_validation_test.go
… internal/wasm/function_definition.go
… internal/wasm/function_definition_test.go
… internal/wasm/global_test.go