This changes wasm.Module's fields, make most of them stored as a slice of values, not ptrs.
As a consequence, this improves not only compilation but also the instantiation perf.

```
$ benchstat aa_old.txt new.txt
name                                    old time/op    new time/op    delta
Initialization/interpreter-32             51.1µs ± 1%    39.3µs ± 4%  -23.11%  (p=0.000 n=10+9)
Initialization/interpreter-multiple-32    37.4µs ± 6%    37.8µs ± 5%     ~     (p=0.579 n=10+10)
Initialization/compiler-32                32.0µs ± 2%    30.1µs ± 1%   -5.88%  (p=0.000 n=10+9)
Initialization/compiler-multiple-32       24.8µs ± 5%    24.6µs ± 5%     ~     (p=0.684 n=10+10)
Compilation/with_extern_cache-32           206µs ± 1%     205µs ± 1%     ~     (p=0.247 n=10+10)
Compilation/without_extern_cache-32       6.12ms ± 1%    6.09ms ± 1%     ~     (p=0.089 n=10+10)

name                                    old alloc/op   new alloc/op   delta
Initialization/interpreter-32              137kB ± 0%     136kB ± 0%   -0.31%  (p=0.000 n=10+10)
Initialization/interpreter-multiple-32     137kB ± 0%     137kB ± 0%     ~     (p=0.134 n=9+8)
Initialization/compiler-32                 141kB ± 0%     137kB ± 0%   -3.07%  (p=0.000 n=9+9)
Initialization/compiler-multiple-32        142kB ± 0%     142kB ± 0%     ~     (p=0.518 n=8+9)
Compilation/with_extern_cache-32          55.7kB ± 0%    55.5kB ± 0%   -0.23%  (p=0.000 n=9+10)
Compilation/without_extern_cache-32       2.00MB ± 0%    2.00MB ± 0%     ~     (p=0.053 n=9+10)

name                                    old allocs/op  new allocs/op  delta
Initialization/interpreter-32               51.0 ± 0%      38.0 ± 0%  -25.49%  (p=0.000 n=10+10)
Initialization/interpreter-multiple-32      57.0 ± 0%      57.0 ± 0%     ~     (all equal)
Initialization/compiler-32                  41.0 ± 0%      38.0 ± 0%   -7.32%  (p=0.000 n=10+10)
Initialization/compiler-multiple-32         47.0 ± 0%      47.0 ± 0%     ~     (all equal)
Compilation/with_extern_cache-32           1.10k ± 0%     1.08k ± 0%   -1.99%  (p=0.000 n=10+10)
Compilation/without_extern_cache-32        33.6k ± 0%     33.6k ± 0%   -0.06%  (p=0.000 n=7+10)
```